NUTCH-1806 Delegate processing of URL domains to crawler commons NUTCH-1942 Remove TopLevelDomain - update DomainStatistics, TLDIndexingFilter and domain URL filters   to use the updated methods in URLUtil - remove TLDScoringFilter - remove package org.apache.nutch.util.domain - move DomainStatistics to org.apache.nutch.util - remove configuration files of domain utils
