+++ default.properties
@@ -130,7 +130,6 @@ plugins.scoring=\
    org.apache.nutch.scoring.opic*:\
    org.apache.nutch.scoring.orphan*:\
    org.apache.nutch.scoring.similarity*:\
-   org.apache.nutch.scoring.tld*:\
    org.apache.nutch.scoring.urlmeta*\
    org.apache.nutch.scoring.metadata*
