+++ src/bin/nutch
@@ -269,7 +269,7 @@ elif [ "$COMMAND" = "filterchecker" ] ; then
 elif [ "$COMMAND" = "normalizerchecker" ] ; then
   CLASS=org.apache.nutch.net.URLNormalizerChecker
 elif [ "$COMMAND" = "domainstats" ] ; then 
-  CLASS=org.apache.nutch.util.domain.DomainStatistics
+  CLASS=org.apache.nutch.util.DomainStatistics
 elif [ "$COMMAND" = "protocolstats" ] ; then
    CLASS=org.apache.nutch.util.ProtocolStatusStatistics
 elif [ "$COMMAND" = "crawlcomplete" ] ; then
