--- src/plugin/tld/src/java/org/apache/nutch/scoring/tld/TLDScoringFilter.java
@@ -1,60 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.nutch.scoring.tld;
-
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.io.Text;
-import org.apache.nutch.crawl.CrawlDatum;
-import org.apache.nutch.crawl.Inlinks;
-import org.apache.nutch.indexer.NutchDocument;
-import org.apache.nutch.indexer.NutchField;
-import org.apache.nutch.parse.Parse;
-import org.apache.nutch.scoring.AbstractScoringFilter;
-import org.apache.nutch.scoring.ScoringFilterException;
-import org.apache.nutch.util.domain.DomainSuffix;
-import org.apache.nutch.util.domain.DomainSuffixes;
-
-/**
- * Scoring filter to boost top-level domains (TLDs).
- */
-public class TLDScoringFilter extends AbstractScoringFilter {
-
-  private DomainSuffixes tldEntries;
-
-  public TLDScoringFilter() {
-    tldEntries = DomainSuffixes.getInstance();
-  }
-
-  @Override
-  public float indexerScore(Text url, NutchDocument doc, CrawlDatum dbDatum,
-      CrawlDatum fetchDatum, Parse parse, Inlinks inlinks, float initScore)
-      throws ScoringFilterException {
-
-    NutchField tlds = doc.getField("tld");
-    float boost = 1.0f;
-
-    if (tlds != null) {
-      for (Object tld : tlds.getValues()) {
-        DomainSuffix entry = tldEntries.get(tld.toString());
-        if (entry != null)
-          boost *= entry.getBoost();
-      }
-    }
-    return initScore * boost;
-  }
-
-}
